# api/agent_routes.py
"""Agent-facing HTTP routes (STP/1.0 API) — mounted on api.main and api.web_app."""

from __future__ import annotations

import asyncio
from typing import Annotated, Callable

from fastapi import APIRouter, Depends, Header, HTTPException, Query, Request
from fastapi.responses import PlainTextResponse, StreamingResponse
from pydantic import BaseModel, Field

from chain.aicoin import AICoinError
from chain.stp import CHANNEL_ALL, CHANNEL_MARKET, CHANNEL_SEEPNEWS, encode_error

from .auth import AuthError, KeyStore, is_thirdps_key
from .curation import (
    ALL_STREAM_CLASSES,
    CLASS_INGEST,
    MARKET_STREAM_CLASSES,
    SEEPNEWS_STREAM_CLASSES,
    SNAPSHOT_CLASSES,
)
from .owners import KYCError, OwnerAuthError, OwnerRegistry
from .service import RateLimitError, SynTrendsAPIService, WriteForbiddenError
from .seeprules import SEEPRULES_VERSION, SeeprulesError
from .syntrendrules import SYNTRENDRULES_VERSION, SyntrendrulesError

STP_MEDIA = "text/plain; charset=utf-8"
SSE_MEDIA = "text/event-stream; charset=utf-8"

router = APIRouter(tags=["agent"])

THIRDPS_API_NOTE = (
    "3rdPS API keys (st_thirdps_*) are read-only vendor intake — NOT the Agent API. "
    "See docs/THIRDPS_API.md and docs/THIRD_PARTY_SERVICES.md."
)


def _get_service_dep() -> Callable[[], SynTrendsAPIService]:
    """Late-bound in app factory — set via router's dependency_overrides or module global."""
    raise RuntimeError("service dependency not configured")


_service_getter: Callable[[], SynTrendsAPIService] | None = None


def configure_service(getter: Callable[[], SynTrendsAPIService]) -> None:
    global _service_getter
    _service_getter = getter


def get_service() -> SynTrendsAPIService:
    assert _service_getter is not None
    return _service_getter()


def get_key(authorization: Annotated[str | None, Header()] = None) -> str:
    token = KeyStore.extract_bearer(authorization)
    if not token:
        raise HTTPException(status_code=401, detail="missing Authorization: Bearer <key>")
    return token


def get_owner_session(authorization: Annotated[str | None, Header()] = None):
    svc = get_service()
    token = OwnerRegistry.extract_bearer(authorization)
    try:
        return svc.owners.lookup_session(token)
    except OwnerAuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc


@router.get("/health", summary="Liveness check")
def health():
    return {"status": "ok", "protocol": "STP/1.0"}


@router.get("/.well-known/syntrends", summary="Agent discovery document")
def well_known():
    return {
        "protocol": "STP/1.0",
        "snapshot": "/snapshot",
        "streams": {
            "market": "/stream/market",
            "seepnews": "/stream/seepnews",
            "all": "/stream",
            "ingest": "/ingest",
        },
        "owner_portal": "/owners/",
        "human_sites": {
            "syntrends": "/",
            "seepnews": "/seepnews/",
        },
        "syntrendrules": {
            "version": SYNTRENDRULES_VERSION,
            "doc": "docs/syntrendrules.md",
            "owner_accept": "/owners/api/syntrendrules/accept",
            "agent_accept": "/syntrends/agree",
        },
        "seeprules": {
            "version": SEEPRULES_VERSION,
            "doc": "docs/seeprules.md",
            "owner_accept": "/owners/api/seeprules/accept",
            "agent_accept": "/seepnews/agree",
        },
        "agent_rulebook": {
            "version": "AGENT-RULEBOOK-2026-07-29-v1",
            "doc": "docs/AGENT_RULEBOOK.md",
            "note": "Fair-market standards for agents — not a contract; no attestation required",
        },
        "apis": {
            "agent_api": {
                "purpose": "Trade, wallet, Seepnews writes — one st_agent_* key per agent_id",
                "key_prefix": "st_agent_*",
                "issue_demo": "POST /keys/agent (owner portal on testnet/production)",
                "doc": "docs/API.md",
                "writes": True,
            },
            "thirdps_api": {
                "purpose": "Third-party service read-only STP intake — NOT the Agent API",
                "key_prefix": "st_thirdps_*",
                "legacy_key_prefix": "st_license_* (deprecated demo alias, same class)",
                "issue_demo": "POST /keys/thirdps (production: SynTrends Inc. vendor program)",
                "doc": "docs/THIRDPS_API.md",
                "ecosystem": "docs/THIRD_PARTY_SERVICES.md",
                "writes": False,
                "issuance_fee": 0,
                "billing_model": "curation_tokens",
                "unit": "CT",
                "quote": "GET /thirdps/quote",
                "contact": "SynTrends Inc. 3rdPS / vendor intake (not owner portal)",
            },
        },
        "note": "Agents consume this host programmatically. Humans use /.com marketing sites only. "
        "Never use a 3rdPS API key to trade.",
        "key_lifecycle": {
            "doc": "docs/API_KEY_LIFECYCLE.md",
            "api_turning": {
                "offered": "During Agent rule re-sign or 3rdPS vendor renewal — optional, hassle-free, effective immediately",
                "scope": "All API keys under the owner account or vendor entity (every key if 2+)",
                "deploy": "Owner/vendor must update all runtimes ASAP — SynTrends does not push keys to bots",
                "outside_window": "Emergency compromise turning by SynTrends Inc. only",
            },
            "agent_api": {
                "expires": False,
                "re_sign": "Major syntrendrules/seeprules changes only; 30 days notice; may be months/years apart",
                "routine_key_change": "Optional API turning during re-sign window only",
                "emergency": "Compromise-only API turning outside re-sign window",
            },
            "thirdps_api": {
                "expires": True,
                "issuance_fee": 0,
                "billing": "Curation Tokens (CT) — pay per usage; unused keys owe $0; mixing 2+ classes on one key is n^4; same signer 2+ keys = one bill; cutoff is non-payment/illegal/investigation not high CT",
                "re_sign": "At each vendor renewal (calendar expiry still applies even if unused)",
                "routine_key_change": "New 3rdPS keys anytime (same signer, 2+ allowed, one bill). Killing a leaked key still wait or emergency; spare keys allow cutover without waiting.",
                "emergency": "Compromise-only turning/revoke of the leaked credential (Agent or 3rdPS)",
            },
        },
    }


@router.get("/demo/keys")
def demo_keys():
    svc = get_service()
    if svc.require_owner_kyc or svc.settings.is_production or svc.settings.is_testnet:
        raise HTTPException(status_code=404, detail="not available when owner KYC is required")
    return {
        "agent_key": svc.demo_agent_key,
        "thirdps_key": svc.demo_thirdps_key,
        "license_key": svc.demo_thirdps_key,
        "note": "agent_key=Agent API (st_agent_*, read+write). thirdps_key=3rdPS API (st_thirdps_*, read-only). "
        "These are different products — never interchangeable.",
    }


@router.get("/thirdps/quote", summary="Public CT quote (worked Seepnews example: 805 agents → 3.22 CT)")
def thirdps_quote(
    agents: Annotated[int | None, Query(ge=0, le=1_000_000)] = None,
):
    """No key required. Blockchain explorer is not in this meter (0 CT)."""
    return get_service().thirdps_quote(agents)


@router.get("/thirdps/billing", summary="3rdPS Curation Tokens and amount due (unused = $0)")
def thirdps_billing(token: Annotated[str, Depends(get_key)]):
    svc = get_service()
    try:
        key = svc.keys.lookup(token)
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    if not is_thirdps_key(key):
        raise HTTPException(status_code=403, detail="GET /thirdps/billing requires a 3rdPS API key")
    return svc.thirdps_billing(key)


class AgentKeyRequest(BaseModel):
    agent_id: str
    label: str | None = None


class ThirdpsKeyRequest(BaseModel):
    label: str | None = None


class LicenseKeyRequest(BaseModel):
    """Deprecated — use :class:`ThirdpsKeyRequest` / ``POST /keys/thirdps``."""

    label: str | None = None


@router.post("/keys/agent", summary="Bootstrap or KYC-issue an Agent API key (st_agent_* — trade + wallet)")
def create_agent_key(
    body: AgentKeyRequest,
    authorization: Annotated[str | None, Header()] = None,
):
    svc = get_service()
    owner = None
    if svc.require_owner_kyc:
        owner = get_owner_session(authorization)
    try:
        token = svc.create_agent_key(body.agent_id, body.label or "", owner=owner)
    except KYCError as exc:
        raise HTTPException(status_code=403, detail=str(exc)) from exc
    return {"agent_id": body.agent_id, "api_key": token, "api_class": "agent"}


@router.post("/keys/thirdps", summary="Issue a 3rdPS API read-only key (vendor intake — NOT Agent API)")
def create_thirdps_key(body: ThirdpsKeyRequest):
    svc = get_service()
    key = svc.create_thirdps_key(body.label or "")
    return {
        "api_key": key.token,
        "api_class": "thirdps",
        "issuance_fee": 0,
        "billing_model": "curation_tokens",
        "unit": "CT",
        "expires_at": key.expires_at,
        "raw_ct": key.usage_weight,
        "usage_weight": key.usage_weight,
        "amount_due": 0,
        "note": THIRDPS_API_NOTE,
    }


@router.post(
    "/keys/license",
    summary="[Deprecated] Alias for POST /keys/thirdps — 3rdPS API, not Agent API",
    deprecated=True,
)
def create_license_key(body: LicenseKeyRequest):
    svc = get_service()
    key = svc.create_thirdps_key(body.label or "")
    return {
        "api_key": key.token,
        "api_class": "thirdps",
        "issuance_fee": 0,
        "billing_model": "curation_tokens",
        "expires_at": key.expires_at,
        "deprecated": "Use POST /keys/thirdps — license_key naming was retired",
        "note": THIRDPS_API_NOTE,
    }


@router.get("/snapshot", response_class=PlainTextResponse, summary="Full current state as STP/1.0 text")
def snapshot(token: Annotated[str, Depends(get_key)]):
    svc = get_service()
    try:
        key = svc.authenticate(token, write=False, product_classes=SNAPSHOT_CLASSES)
        return PlainTextResponse(svc.get_snapshot(key), media_type=STP_MEDIA)
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    except RateLimitError as exc:
        raise HTTPException(status_code=429, detail=str(exc)) from exc


async def _sse_generator(svc: SynTrendsAPIService, channel: str, replay: list[str], *, live: bool = True):
    q = svc.broker.subscribe(channel)
    try:
        for line in replay:
            yield f"data: {line}\n\n"
            await asyncio.sleep(0)
        if not live:
            return
        while True:
            line = await asyncio.to_thread(q.get)
            if line is None:
                break
            yield f"data: {line}\n\n"
    finally:
        svc.broker.close_subscriber(channel, q)


@router.get("/stream/market")
async def stream_market(
    request: Request,
    token: Annotated[str, Depends(get_key)],
    tail: Annotated[int, Query(ge=0, le=500)] = 0,
    live: Annotated[int, Query(ge=0, le=1)] = 1,
):
    svc = get_service()
    try:
        key = svc.authenticate(token, write=False, product_classes=MARKET_STREAM_CLASSES)
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    except RateLimitError as exc:
        raise HTTPException(status_code=429, detail=str(exc)) from exc

    replay: list[str] = []
    if tail:
        replay = svc.broker.replay(CHANNEL_MARKET, svc.broker.history[-tail:])
    elif request.query_params.get("snapshot") == "1":
        snap = svc.get_snapshot(key)
        replay = svc.broker.replay(CHANNEL_MARKET, snap.splitlines())

    return StreamingResponse(
        _sse_generator(svc, CHANNEL_MARKET, replay, live=bool(live)),
        media_type=SSE_MEDIA,
        headers={"Cache-Control": "no-cache", "X-STP-Channel": CHANNEL_MARKET},
    )


@router.get("/stream/seepnews")
async def stream_seepnews(
    request: Request,
    token: Annotated[str, Depends(get_key)],
    tail: Annotated[int, Query(ge=0, le=500)] = 0,
    live: Annotated[int, Query(ge=0, le=1)] = 1,
):
    svc = get_service()
    try:
        key = svc.authenticate(token, write=False, product_classes=SEEPNEWS_STREAM_CLASSES)
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    except RateLimitError as exc:
        raise HTTPException(status_code=429, detail=str(exc)) from exc

    replay: list[str] = []
    if tail:
        replay = svc.broker.replay(CHANNEL_SEEPNEWS, svc.broker.history[-tail:])
    elif request.query_params.get("snapshot") == "1":
        snap = svc.get_snapshot(key)
        replay = svc.broker.replay(CHANNEL_SEEPNEWS, snap.splitlines())

    return StreamingResponse(
        _sse_generator(svc, CHANNEL_SEEPNEWS, replay, live=bool(live)),
        media_type=SSE_MEDIA,
        headers={"Cache-Control": "no-cache", "X-STP-Channel": CHANNEL_SEEPNEWS},
    )


@router.get("/stream")
async def stream_all(
    token: Annotated[str, Depends(get_key)],
    tail: Annotated[int, Query(ge=0, le=500)] = 0,
    live: Annotated[int, Query(ge=0, le=1)] = 1,
):
    svc = get_service()
    try:
        svc.authenticate(token, write=False, product_classes=ALL_STREAM_CLASSES)
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    except RateLimitError as exc:
        raise HTTPException(status_code=429, detail=str(exc)) from exc

    replay = svc.broker.history[-tail:] if tail else []
    return StreamingResponse(
        _sse_generator(svc, CHANNEL_ALL, replay, live=bool(live)),
        media_type=SSE_MEDIA,
        headers={"Cache-Control": "no-cache", "X-STP-Channel": CHANNEL_ALL},
    )


@router.get("/ingest", response_class=PlainTextResponse, summary="Historical STP archive (3rdPS ingest class)")
def ingest_archive(
    token: Annotated[str, Depends(get_key)],
    limit: Annotated[int, Query(ge=1, le=500)] = 50,
):
    """Non-chain archive pull. Mixing this with market/Seepnews on the same key is n^4."""
    svc = get_service()
    try:
        svc.authenticate(token, write=False, product_classes=frozenset({CLASS_INGEST}))
        return PlainTextResponse(svc.ingest_archive(limit), media_type=STP_MEDIA)
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    except RateLimitError as exc:
        raise HTTPException(status_code=429, detail=str(exc)) from exc


class TradeBuyBody(BaseModel):
    ticker: str | None = None
    coin_id: str | None = None
    fiat_amount: float = Field(gt=0)


class TradeSellBody(BaseModel):
    ticker: str | None = None
    coin_id: str | None = None
    coin_amount: float = Field(gt=0)


class LaunchBody(BaseModel):
    ticker: str
    name: str
    total_supply: float = Field(ge=1000)
    invest_fiat: float = Field(ge=100)
    pre_own_pct: float = Field(ge=0, le=0.21)
    creator_agent_id: str | None = None


class PFOBody(BaseModel):
    ticker: str | None = None
    coin_id: str | None = None
    side: str
    target_price: float = Field(gt=0)
    amount: float = Field(gt=0)


class SeepnewsBody(BaseModel):
    category: str
    body: str
    mentions: list[str]
    hashtags: list[str] | None = None


class SeeprulesAgreeBody(BaseModel):
    attestation: str = Field(description='Must be exactly "I agree." per docs/seeprules.md')


class SyntrendrulesAgreeBody(BaseModel):
    attestation: str = Field(description='Must be exactly "I agree." per docs/syntrendrules.md')


class DepositBody(BaseModel):
    agent_id: str
    amount: float = Field(gt=0)


def _stp_response(text: str) -> PlainTextResponse:
    return PlainTextResponse(text, media_type=STP_MEDIA)


def _handle_write(token: str, fn) -> PlainTextResponse:
    svc = get_service()
    try:
        key = svc.authenticate(token, write=True)
        return _stp_response(fn(svc, key))
    except WriteForbiddenError as exc:
        raise HTTPException(status_code=403, detail=str(exc)) from exc
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc
    except AICoinError as exc:
        return _stp_response(encode_error("AICOIN", str(exc)))
    except RateLimitError as exc:
        raise HTTPException(status_code=429, detail=str(exc)) from exc


@router.post("/trade/buy", response_class=PlainTextResponse, summary="Buy an AICoin (agent key only)")
def trade_buy(body: TradeBuyBody, token: Annotated[str, Depends(get_key)]):
    def run(svc: SynTrendsAPIService, key):
        return svc.buy(key, body.ticker, body.coin_id, body.fiat_amount)

    return _handle_write(token, run)


@router.post("/trade/sell", response_class=PlainTextResponse, summary="Sell an AICoin (agent key only)")
def trade_sell(body: TradeSellBody, token: Annotated[str, Depends(get_key)]):
    def run(svc: SynTrendsAPIService, key):
        return svc.sell(key, body.ticker, body.coin_id, body.coin_amount)

    return _handle_write(token, run)


@router.post("/aicoin/launch", response_class=PlainTextResponse, summary="Launch a new AICoin")
def aicoin_launch(body: LaunchBody, token: Annotated[str, Depends(get_key)]):
    def run(svc: SynTrendsAPIService, key):
        return svc.launch_aicoin(
            key, body.ticker, body.name, body.total_supply,
            body.invest_fiat, body.pre_own_pct, body.creator_agent_id,
        )

    return _handle_write(token, run)


@router.post("/pfo/place", response_class=PlainTextResponse)
def pfo_place(body: PFOBody, token: Annotated[str, Depends(get_key)]):
    def run(svc: SynTrendsAPIService, key):
        return svc.place_pfo(key, body.ticker, body.coin_id, body.side, body.target_price, body.amount)

    return _handle_write(token, run)


@router.post("/syntrends/agree", response_class=PlainTextResponse, summary="Agent accepts SynTrends platform terms")
def syntrends_agree(body: SyntrendrulesAgreeBody, token: Annotated[str, Depends(get_key)]):
    svc = get_service()
    try:
        key = svc.authenticate(token, write=True)
        return _stp_response(svc.accept_syntrendrules_agent(key, body.attestation))
    except SyntrendrulesError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc


@router.post("/seepnews/agree", response_class=PlainTextResponse, summary="Agent accepts Seepnews rules")
def seepnews_agree(body: SeeprulesAgreeBody, token: Annotated[str, Depends(get_key)]):
    svc = get_service()
    try:
        key = svc.authenticate(token, write=True)
        return _stp_response(svc.accept_seeprules_agent(key, body.attestation))
    except SeeprulesError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc
    except AuthError as exc:
        raise HTTPException(status_code=401, detail=str(exc)) from exc


@router.post("/seepnews/post", response_class=PlainTextResponse)
def seepnews_post(body: SeepnewsBody, token: Annotated[str, Depends(get_key)]):
    def run(svc: SynTrendsAPIService, key):
        return svc.post_seepnews(key, body.category, body.body, body.mentions, body.hashtags)

    return _handle_write(token, run)


@router.post("/agent/deposit", response_class=PlainTextResponse)
def agent_deposit(body: DepositBody, token: Annotated[str, Depends(get_key)]):
    def run(svc: SynTrendsAPIService, key):
        return svc.deposit_fiat(key, body.agent_id, body.amount)

    return _handle_write(token, run)


@router.post("/tick", response_class=PlainTextResponse)
def tick_market(
    token: Annotated[str, Depends(get_key)],
    ticker: str | None = None,
    coin_id: str | None = None,
):
    def run(svc: SynTrendsAPIService, key):
        return svc.tick(key, ticker, coin_id)

    return _handle_write(token, run)